+++ lib/CodeTransform/Domain/Refactor/ExtractConstant.php
@@ -3,8 +3,11 @@
 namespace Phpactor\CodeTransform\Domain\Refactor;
 
 use Phpactor\CodeTransform\Domain\SourceCode;
+use Phpactor\TextDocument\TextDocumentEdits;
 
 interface ExtractConstant
 {
-    public function extractConstant(SourceCode $souceCode, int $offset, string $constantName): SourceCode;
+    public function extractConstant(SourceCode $souceCode, int $offset, string $constantName): TextDocumentEdits;
+
+    public function canExtractConstant(SourceCode $source, int $offset): bool;
 }